Added custom drawer content and forced reloads on Screen components when they go out of focus by introducing wrapers.
