+++ apps/llm/app/context.ts
@@ -1,0 +1,5 @@
+import { createContext } from 'react';
+
+export const GeneratingContext = createContext({
+  setGlobalGenerating: (_newState: boolean) => {},
+});
